Refactor osal directory and create separated common content
