+++ osal/linux/osal.c
